% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nichenet.R
\name{Clone_expr}
\alias{Clone_expr}
\title{Get clone expression matrix, for each gene use mean value of all cells of the same clone}
\usage{
Clone_expr(scObject, nnt, seurat_layer = "counts")
}
\arguments{
\item{scObject}{Seurat, input seurat object in which tag expression matrix is merged in gene expression matrix}

\item{nnt}{a list contains connectome information of all niches}

\item{seurat_layer}{layer in scObject used to calculate expression level of genes}
}
\value{
clone expression matrix of all nontag genes
}
\description{
Get clone expression matrix, for each gene use mean value of all cells of the same clone
}
\examples{
clonematrix=Clone_expr(scObject,nnt,seurat_layer="counts")
}
