% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{codesplit}
\alias{codesplit}
\title{Translate code to tag and cell type}
\usage{
codesplit(code, code2tag, code2type)
}
\arguments{
\item{code}{code of cells}

\item{code2tag}{each position of code corresponds to a tag}

\item{code2type}{each position of code corresponds to a celltype}
}
\value{
text of tag and celltype
}
\description{
Translate code to tag and cell type
}
\examples{
codesplot()
}
