% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nichenet.R
\name{Dnichenetwork}
\alias{Dnichenetwork}
\title{Calculate niche-niche connections with connectome data from nichetag method}
\usage{
Dnichenetwork(
  tag_expression,
  cell_clusters,
  share_method = "min",
  direction = FALSE
)
}
\arguments{
\item{tag_expression}{dataframe, input tag expression matrix in which cells are rownames and tag expression are colnames}

\item{cell_clusters}{factor, input clusters of all cells in tag_expression matrix}

\item{share_method}{min,max,mean, mothed to calculate connection strength between different clones}

\item{direction}{TorF, if or not to distinguish tag-sender or tag-receiver}
}
\value{
a list contains connectome information of all niches
}
\description{
This fuction use tag expression matrix to infer the connection among niches labeled with different nichetags
}
\examples{
nnt=Dnichenetwork(tag_expression,cell_clusters)
summary(nnt)
nnt=nichenetwork(tag_expression,cell_clusters,share_method="mean")
dnnt=Dnichenetwork(tag_expression,cell_clusters,direction = T)
}
