% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nichenet.R
\name{findoverlap}
\alias{findoverlap}
\title{find overlap of clones}
\usage{
findoverlap(clonetag, share_method = "min")
}
\arguments{
\item{clonetag}{clones}

\item{share_method}{min,max,mean, mothed to calculate connection strength between different clones}
}
\value{
overlap
}
\description{
find overlap of clones
}
\examples{
findoverlap()
}
