% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_niche.R
\name{print_clustertag}
\alias{print_clustertag}
\title{print cluster vs tag number and types}
\usage{
print_clustertag(nnt, file = "clustertag.pdf")
}
\arguments{
\item{nnt}{a list contains connectome information of all niches, the result of Dnichenetwork}

\item{file}{pdf}
}
\value{
pdf
}
\description{
print cluster vs tag number and types
}
\examples{
print_clustertag(nnt)
}
