% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_niche.R
\name{print_nichenet}
\alias{print_nichenet}
\title{print niche network}
\usage{
print_nichenet(
  nnt,
  file = "nichenet.pdf",
  vsize = 1,
  esize = 1,
  vertex.label.cex = 1,
  axes = FALSE,
  width = 7,
  height = 7,
  weighted = FALSE,
  direction = FALSE
)
}
\arguments{
\item{nnt}{a list contains connectome information of all niches, the result of Dnichenetwork}

\item{file}{file names to be saved}

\item{vsize}{vertex size}

\item{esize}{edge size}

\item{vertex.label.cex}{vertex label size}

\item{axes}{TorF, if or not draw axes}

\item{width}{file size}

\item{height}{file size}

\item{weighted}{layout caculation parameter}

\item{direction}{layout caculation parameter}
}
\value{
pdf file
}
\description{
print niche network
}
\examples{
print_nichenet(nnt,file="nichenet-w2.pdf",vsize = 10)
}
