% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_niche.R
\name{print_nichetag}
\alias{print_nichetag}
\title{print tag vs cell number and types}
\usage{
print_nichetag(nnt, file = "nichetag.pdf")
}
\arguments{
\item{nnt}{a list contains connectome information of all niches, the result of Dnichenetwork}

\item{file}{pdf}
}
\value{
pdf
}
\description{
print tag vs cell number and types
}
\examples{
print_nichetag(nnt)
}
