% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{quality_control}
\alias{quality_control}
\title{Title}
\usage{
quality_control(tag_expression)
}
\arguments{
\item{tag_expression}{dataframe, input tag expression matrix in which cells are rownames and tag expression are colnames}
}
\value{
dataframe, removed uncorrected tag
}
\description{
Title
}
\examples{
tag_expression=quality_control(tag_expression)
}
